Set GC_TTL to 14 days
This environment variable will be used as the value of the -gc-ttl flag, which
determines how long Autojoin will leave a DNS record in place when it hasn't
heard from a client. The default value is 3 hours, which is too aggressive,
especially in light of M-Lab wanting to leverage Autojoin for M-Lab-managed
VMs.

Additionally, I added the instance_tag field under network so that AppEngine
will assign that network tag to any created instances. We will use that tag in
creating a new firewall rule which opens port 9090 on Autojoin instance.
Prometheus needs this port to be open in order to scrape the AppEngine
instances.