Sets GC_TTL=14d + adds new metrics package (#62)
* Set GC_TTL to 14 days

This environment variable will be used as the value of the -gc-ttl flag, which
determines how long Autojoin will leave a DNS record in place when it hasn't
heard from a client. The default value is 3 hours, which is too aggressive,
especially in light of M-Lab wanting to leverage Autojoin for M-Lab-managed
VMs.

Additionally, I added the instance_tag field under network so that AppEngine
will assign that network tag to any created instances. We will use that tag in
creating a new firewall rule which opens port 9090 on Autojoin instance.
Prometheus needs this port to be open in order to scrape the AppEngine
instances.

* Adds new metrics package with DNSExpiration metric

This changes adds a new metrics package where a new metric named DNSExpiration
is added. This metric will track the date on which the DNS record for a given
hostname will be removed from Cloud DNS. This change also moves the previous
RequestHandlerDuration metric to this new package.