Replaces all dot with dashes in service account names
This will all us to have Autojoin organizations names with dots. Specifically,
it would be nice to be able to use an ISP's primary domain name as the org name
in the Autojoin system. The domain name is self-documenting, since you can just
paste it into a browser to get more information about the org. Additionally it
is guaranteed to be globally unique.