Reduces GC_TTL from 14d to 1d
The TTL was originally only 3h, but we bumped it up to 14d in anticipation of
leveraging the Autojoin system from M-Lab-managed VMs to give us plenty of
warning time before a bug might delete all the DNS records for a substantial
number of production VMs. However, we never implemented the M-Lab-managed VMs
using the Autojoin system.

This commit lowers the GC_TTL down to 1d, which should be plenty, and is in any
case 8 times longer than the original 3h setting.

When this value is too high for no useful reason, it causes the Autojoin system
to keep a bunch of records around for Autojoin machines that may no longer
actually exist, which ends up polluting monitoring and observability.